move the real tests to their own test file
